c++11: provide custom auto_ptr implementation in namespace msgpack
The std::auto_ptr is deprecated since c++11. To support the new
standard and get rid of the deprecation warnings we add a custom
msgpack::auto_ptr. This wrapper exposes the same interface as the
deprecated std::auto_ptr. In fact, if compiling under c++98 it
falls back to the original STL std::auto_ptr. However, compiling
with the new standard c++11 uses a wrapped std::unique_ptr (with
hidden move semantics) instead.
Note: this wrapper suffers from the same problems that led to the
deprecation of std::auto_ptr!